+++ app/Http/Controllers/LabTestController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Lab_test;
+use App\Models\LabTest;
 use App\Http\Requests\StoreLab_testRequest;
 use App\Http\Requests\UpdateLab_testRequest;
 use App\Services\LabTestService;
@@ -12,14 +12,15 @@ use Symfony\Component\HttpFoundation\Response;
 class LabTestController extends Controller
 {
 
-   protected $labTestService;
-   public function __construct(LabTestService $labTestService)
+    protected LabTestService $labTestService;
+    public function __construct(LabTestService $labTestService)
     {
         $this->labTestService = $labTestService;
     }
 
     public function index(): JsonResponse
     {
+        $this->authorize('viewAny', LabTest::class);
         $tests = $this->labTestService->getAllTests();
         return response()->json([
             'success' => true,
@@ -29,6 +30,7 @@ class LabTestController extends Controller
 
     public function store(StoreLab_testRequest $request): JsonResponse
     {
+        $this->authorize('create', LabTest::class);
         $test = $this->labTestService->createTest($request->validated());
         return response()->json([
             'success' => true,
@@ -40,6 +42,7 @@ class LabTestController extends Controller
     public function show(int $id): JsonResponse
     {
         $test = $this->labTestService->getTestById($id);
+        $this->authorize('view', $test);
         return response()->json([
             'success' => true,
             'data'    => $test
@@ -48,6 +51,8 @@ class LabTestController extends Controller
 
     public function update(UpdateLab_testRequest $request, int $id): JsonResponse
     {
+        $test = $this->labTestService->getTestById($id);
+        $this->authorize('update', $test);
         $this->labTestService->updateTest($id, $request->validated());
         return response()->json([
             'success' => true,
@@ -57,6 +62,8 @@ class LabTestController extends Controller
 
     public function destroy(int $id): JsonResponse
     {
+        $test = $this->labTestService->getTestById($id);
+        $this->authorize('delete', $test);
         $this->labTestService->deleteTest($id);
         return response()->json([
             'success' => true,
