+++ app/Http/Controllers/MedicalConditionController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreMedicalConditionRequest;
 use App\Http\Requests\UpdateMedicalConditionRequest;
+use App\Models\MedicalCondition;
 use App\Services\MedicalConditionService;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
+
 class MedicalConditionController extends Controller
 {
 
@@ -18,15 +20,17 @@ class MedicalConditionController extends Controller
     }
     public function index(): JsonResponse
     {
+        $this->authorize('viewAny', MedicalCondition::class);
         $medical_conditions = $this->medicalConditionService->getAllMedicalCondition();
         return response()->json(['success' => true, 'data' => $medical_conditions], Response::HTTP_OK);
     }
 
 
 
- 
+
     public function store(StoreMedicalConditionRequest $request): JsonResponse
     {
+        $this->authorize('create', MedicalCondition::class);
         $medical_condition = $this->medicalConditionService->createMedicalCondition($request->validated());
         return response()->json([
             'success' => true,
@@ -39,6 +43,7 @@ class MedicalConditionController extends Controller
     public function show(int $id): JsonResponse
     {
         $medical_condition = $this->medicalConditionService->getMedicalConditionById($id);
+        $this->authorize('view', $medical_condition);
         return response()->json(['success' => true, 'data' => $medical_condition], Response::HTTP_OK);
     }
 
@@ -47,6 +52,8 @@ class MedicalConditionController extends Controller
 
     public function update(UpdateMedicalConditionRequest $request, int $id): JsonResponse
     {
+        $medical_condition = $this->medicalConditionService->getMedicalConditionById($id);
+        $this->authorize('update', $medical_condition);
         $this->medicalConditionService->updateMedicalCondition($id, $request->validated());
         return response()->json([
             'success' => true,
@@ -56,6 +63,8 @@ class MedicalConditionController extends Controller
 
     public function destroy(int $id): JsonResponse
     {
+        $medical_condition = $this->medicalConditionService->getMedicalConditionById($id);
+        $this->authorize('delete', $medical_condition);
         $this->medicalConditionService->deleteMedicalCondition($id);
         return response()->json([
             'success' => true,
